+++ com/hackbulgaria/ddsystem/models/Coordinates.java
@@ -1,15 +1,15 @@
 package com.hackbulgaria.ddsystem.models;
 
 public class Coordinates {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Coordinates(int x, int y) {
+    public Coordinates(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public static int distance(Coordinates a, Coordinates b) {
+    public static double distance(Coordinates a, Coordinates b) {
         return (a.x - b.x) * (a.x - b.x) +
                 (a.y - b.y) * (a.y - b.y);
     }
